Move new EPG files out of Epg directory
