--- lib/python/Components/Epg/Makefile.am
@@ -1,3 +1,0 @@
-installdir = $(pkglibdir)/python/Components/Epg
-
-install_PYTHON = *.py
