--- lib/python/Screens/Epg/Makefile.am
@@ -1,4 +1,0 @@
-installdir = $(pkglibdir)/python/Screens/Epg
-
-install_PYTHON = *.py
-
